Split packet manager into network manager and packet manager, implement packet manager family
